Add enterprise-v113 and enterprise-php54-v113
- Introduce test suites for 'EE with Magento v1.13' and 'EE with Magento v1.13 and PHP 5.4'
- Deploy against v1.13-php54 tag for EE, from git, that includes the requisite patches for PHP 54 support in Magento
- Add more cache-cleaning power to the scripts that clean cache as part of the test suite, to make them work under EE v1.13
- Even more de-duplication of test suites, separating php shared examples out into a separate file
- Ensure all shared tests are in all suites
- Verify Zend Opcache is in use in `php -v` output
