+++ .kitchen.yml
@@ -39,6 +39,9 @@ platforms:
     run_list:
     - recipe[yum]
 
+# NB: Every build will specify 'community' or 'enterprise' edition, but all
+# other test suite attributes are assumed to be php55 and git with a single
+# redis instance locally on the same server, when not specified.
 suites:
   - name: community
     provisioner:
@@ -90,6 +93,7 @@ suites:
         flavor: 'enterprise'
         install_method: 'git'
         git_repository: 'git@github.com:racker/enterprise-magento-chef-demo.git'
+        git_revision: 'v1.14.0.1'
         git_deploykey: <%= ENV['RACKSPACE_MAGENTOSTACK_DEPLOYKEY'] %>
 
   - name: enterprise-php54
@@ -104,6 +108,39 @@ suites:
         flavor: 'enterprise'
         install_method: 'git'
         git_repository: 'git@github.com:racker/enterprise-magento-chef-demo.git'
+        git_revision: 'v1.14.0.1'
+        git_deploykey: <%= ENV['RACKSPACE_MAGENTOSTACK_DEPLOYKEY'] %>
+        php:
+          version: 'php54'
+
+  - name: enterprise-v113
+    run_list:
+      - recipe[wrapper::default]
+    attributes:
+      rackspace:
+        cloud_credentials:
+          username: 'rackops'
+          api_key: <%= ENV['RACKSPACE_API_KEY'] %>
+      magentostack:
+        flavor: 'enterprise'
+        install_method: 'git'
+        git_repository: 'git@github.com:racker/enterprise-magento-chef-demo.git'
+        git_revision: 'v1.13.1.0'
+        git_deploykey: <%= ENV['RACKSPACE_MAGENTOSTACK_DEPLOYKEY'] %>
+
+  - name: enterprise-php54-v113
+    run_list:
+      - recipe[wrapper::default]
+    attributes:
+      rackspace:
+        cloud_credentials:
+          username: 'rackops'
+          api_key: <%= ENV['RACKSPACE_API_KEY'] %>
+      magentostack:
+        flavor: 'enterprise'
+        install_method: 'git'
+        git_repository: 'git@github.com:racker/enterprise-magento-chef-demo.git'
+        git_revision: 'v1.13.1.0-php54'
         git_deploykey: <%= ENV['RACKSPACE_MAGENTOSTACK_DEPLOYKEY'] %>
         php:
           version: 'php54'
