+++ test/integration/community-redis/serverspec/default_spec.rb
@@ -9,3 +9,7 @@ end
 describe 'community magento under mysql with multiple redis' do
   it_behaves_like 'magento under mysql'
 end
+
+describe 'community magento with php55' do
+  it_behaves_like 'php55 under apache'
+end
