+++ test/integration/community/serverspec/caches_spec.rb
@@ -6,6 +6,7 @@ describe 'community with single redis' do
   before do
     clear_out_redis '-a runstatepasswordsingle -n 0'
     clear_out_redis '-a runstatepasswordsingle -n 2'
+    flush_all_magento_caches
     page_returns
   end
   it_behaves_like 'magento redis cache', '-a runstatepasswordsingle -n 0'
