+++ test/integration/enterprise-cloudfiles/serverspec/default_spec.rb
@@ -7,3 +7,7 @@ end
 describe 'enterprise magento under mysql with single redis' do
   it_behaves_like 'magento under mysql'
 end
+
+describe 'community magento with php55' do
+  it_behaves_like 'php55 under apache'
+end
