+++ test/integration/enterprise-v113/serverspec/default_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-description = 'enterprise magento v1.14 under apache with single redis, mysql, and php 5.4'
+description = 'enterprise magento v1.14 under apache with single redis, mysql, and php 5.5'
 describe description do
   it_behaves_like 'magento under apache'
 end
@@ -10,5 +10,5 @@ describe description do
 end
 
 describe description do
-  it_behaves_like 'php54 under apache'
+  it_behaves_like 'php55 under apache'
 end
