+++ test/integration/helpers/serverspec/spec_helper.rb
@@ -4,12 +4,56 @@ require 'net/http'
 require 'openssl'
 
 require_relative 'default_examples'
+require_relative 'cache_examples'
 require_relative 'apache_examples'
+require_relative 'php_examples'
 require_relative 'mysql_examples'
 require_relative 'magento_admin_examples'
 
 set :backend, :exec
 set :path, '/sbin:/usr/local/sbin:/bin:/usr/bin:$PATH'
+
+def redhat_family_values
+  res = {}
+  res['docroot'] = '/var/www/html/magento'
+  res['apache_service_name'] = 'httpd'
+  res['fpm_service_name'] = 'php-fpm'
+  res['apache2ctl'] = '/usr/sbin/apachectl'
+  res
+end
+
+def debian_family_values
+  res = {}
+  res['docroot'] = '/var/www/magento'
+  res['apache_service_name'] = 'apache2'
+  res['fpm_service_name'] = 'php5-fpm'
+  res['apache2ctl'] = '/usr/sbin/apache2ctl'
+  res
+end
+
+def family_value(str)
+  if os[:family] == 'redhat'
+    redhat_family_values[str]
+  else
+    debian_family_values[str]
+  end
+end
+
+def docroot
+  family_value('docroot')
+end
+
+def apache_service_name
+  family_value('apache_service_name')
+end
+
+def fpm_service_name
+  family_value('fpm_service_name')
+end
+
+def apache2ctl
+  family_value('apache2ctl')
+end
 
 def page_returns(url = 'http://localhost:8080/', host = 'localhost', ssl = false)
   uri = URI.parse(url)
@@ -22,8 +66,4 @@ def page_returns(url = 'http://localhost:8080/', host = 'localhost', ssl = false
   req = Net::HTTP::Get.new(uri.request_uri)
   req.initialize_http_header('Host' => host)
   http.request(req).body
-end
-
-def clear_out_redis(args, redis_path = '/usr/local/bin')
-  "#{redis_path}/redis-cli #{args} --no-raw flushdb"
 end
